login function revised
