+++ src/controllers/authController.ts
@@ -1,6 +1,7 @@
 import { Request, Response } from "express";
 import User from "../models/User";
 import bcrypt from "bcrypt";
+import jwt from "jsonwebtoken";
 
 // POST create new user
 export const createUser = async (
@@ -72,4 +73,56 @@ export const createUser = async (
       error: "Failed to create user",
     });
   }
+};
+
+// Post login user
+export const login = async (req: Request, res: Response) => {
+  const { email, password } = req.body;
+
+  // Validate required fields
+  if (!email || !password) {
+    res.status(400).json({
+      success: false,
+      error: "Please provide all required fields",
+    });
+    return;
+  }
+
+  // Find the user in the database by email
+  const user = await User.findOne({ email: email.toLowerCase() });
+
+  // If user is not found, send an authentication failed response.
+  if (!user) {
+    return res.status(401).json({ message: "Authentication failed" });
+  }
+
+  // Compare the provided password with the stored hashed password
+  const passwordIsValid = await bcrypt.compare(password, user.password);
+
+  // If password is invalid, send an authentication failed response
+  if (!passwordIsValid) {
+    return res
+      .status(401)
+      .json({ message: "Authentication failed passwordIsValid" });
+  }
+
+  // Get the secret key from environment variables
+  const secretKey = process.env.JWT_SECRET_KEY || "";
+
+  // Generate a JWT token for the authenticated user
+  const token = jwt.sign(
+    {
+      id: user._id,
+      name: user.name,
+      email: user.email,
+      isAdmin: user.isAdmin,
+    },
+    secretKey,
+    { expiresIn: "1h" }
+  );
+
+  return res.status(200).json({
+    success: true,
+    token,
+  });
 };
